% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi_colour.R
\name{coloured_print}
\alias{coloured_print}
\title{Add ANSI colour to a string}
\usage{
coloured_print(text, colour = "green")
}
\arguments{
\item{text}{string}

\item{colour}{colour ("red", "green", etc.) to add. Defaults to "green".}
}
\value{
a character string with ANSI codes surrounding it.
}
\description{
Add ANSI colour to a string
}
\keyword{internal}
