% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layouts.R
\name{get_layouts}
\alias{get_layouts}
\title{Get the gridify layouts}
\usage{
get_layouts()
}
\value{
A vector listing out the names of the layout functions from the gridify package.
}
\description{
Lists out all the layout functions exported from the gridify package.
}
\examples{
get_layouts()
}
\seealso{
\code{\link[=complex_layout]{complex_layout()}}, \code{\link[=simple_layout]{simple_layout()}}, \code{\link[=pharma_layout_base]{pharma_layout_base()}},
\code{\link[=pharma_layout_A4]{pharma_layout_A4()}}, \code{\link[=pharma_layout_letter]{pharma_layout_letter()}}
}
