% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\name{gpar_args}
\alias{gpar_args}
\title{Get \code{grid::gpar} arguments}
\usage{
gpar_args(gpar)
}
\arguments{
\item{gpar}{a \code{grid::gpar} object.}
}
\value{
a list.
}
\description{
Get \code{grid::gpar} arguments
}
\keyword{internal}
