% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\name{gpar_call}
\alias{gpar_call}
\title{Convert \code{grid::gpar} to a call}
\usage{
gpar_call(gpar)
}
\arguments{
\item{gpar}{a \code{grid::gpar} object.}
}
\value{
a call.
}
\description{
Convert \code{grid::gpar} to a call
}
\keyword{internal}
