% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_utils.R
\name{grid_unit_type}
\alias{grid_unit_type}
\title{Wrapper for \code{grid::unitType} which supports older R versions}
\usage{
grid_unit_type(x, use_grid = TRUE)
}
\arguments{
\item{x}{a grid::unit}

\item{use_grid}{means try to call grid::unitType if it exists.
The main purpose of this argument is to have full test coverage in tests.
Default TRUE.}
}
\value{
a character vector with unit type for each element.
}
\description{
Wrapper for \code{grid::unitType} which supports older R versions
}
\keyword{internal}
