% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-classes.R
\docType{class}
\name{gridifyCells-class}
\alias{gridifyCells-class}
\title{gridifyCells class}
\description{
Class for creating a list of cells in a gridify layout.
}
\section{Slots}{

\describe{
\item{\code{cells}}{A list of cell objects.}
}}

