% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-classes.R
\name{gridifyCells}
\alias{gridifyCells}
\title{Create a gridifyCells}
\usage{
gridifyCells(...)
}
\arguments{
\item{...}{Arguments passed to the new function to create an instance of the gridifyCells class.
Each argument should be the result of a call to gridifyCell.}
}
\value{
An instance of the gridifyCells class.
}
\description{
Function for creating a new instance of the gridifyCells class.
gridifyCells consists of multiple gridifyCell objects and is an input object for gridifyLayout.
}
\examples{
cell1 <- gridifyCell(
  row = 1,
  col = 1,
  x = 0.5,
  y = 0.5,
  hjust = 0.5,
  vjust = 0.5,
  rot = 0,
  gpar = grid::gpar()
)
cell2 <- gridifyCell(
  row = 2,
  col = 2,
  x = 0.5,
  y = 0.5,
  hjust = 0.5,
  vjust = 0.5,
  rot = 0,
  gpar = grid::gpar()
)
cells <- gridifyCells(title = cell1, footer = cell2)
}
\seealso{
\code{\link[=gridifyLayout]{gridifyLayout()}}
}
