% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-classes.R
\docType{class}
\name{gridifyClass-class}
\alias{gridifyClass-class}
\title{gridifyClass class}
\description{
Class for creating a gridify object.
}
\section{Slots}{

\describe{
\item{\code{object}}{A grob like object.}

\item{\code{layout}}{A gridifyLayout object.}

\item{\code{elements}}{A list of text elements, calls to \code{set_cell()}.}
}}

