% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-classes.R
\docType{class}
\name{gridifyLayout-class}
\alias{gridifyLayout-class}
\title{gridifyLayout class}
\description{
Class for creating a layout for a gridify object.
}
\section{Slots}{

\describe{
\item{\code{nrow}}{An integer specifying the number of rows in the layout.}

\item{\code{ncol}}{An integer specifying the number of columns in the layout.}

\item{\code{heights}}{A \code{grid::unit()} call specifying the heights of the rows.}

\item{\code{widths}}{A \code{grid::unit()} call specifying the widths of the columns.}

\item{\code{margin}}{A \code{grid::unit()} specifying the margins around the object.}

\item{\code{global_gpar}}{A \code{grid::gpar()} object specifying the global graphical parameters.}

\item{\code{background}}{A string with background colour.}

\item{\code{adjust_height}}{A logical value indicating whether to adjust the height of the object.
Only applies for cells with height defined in cm, mm, inch or lines units.}

\item{\code{object}}{A grob object.}

\item{\code{cells}}{A list of cell objects.}
}}

