% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-utils.R
\name{has_metadata_payload}
\alias{has_metadata_payload}
\title{Check whether a metadata payload contains values}
\usage{
has_metadata_payload(payload)
}
\arguments{
\item{payload}{A metadata payload.}
}
\value{
\code{TRUE} when the payload contains at least one metadata value.
}
\description{
Check whether a metadata payload contains values
}
\keyword{internal}
