% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-methods.R
\name{show,gridifyLayout-method}
\alias{show,gridifyLayout-method}
\title{Show method for gridifyLayout}
\usage{
\S4method{show}{gridifyLayout}(object)
}
\arguments{
\item{object}{A gridifyLayout object.}
}
\value{
The list of cells defined in the layout.
}
\description{
Method for showing a gridifyLayout object. It prints the names of the cells in the layout.
}
\examples{
show(complex_layout())

}
\seealso{
\code{\link[=gridifyLayout]{gridifyLayout()}}
}
