% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-classes.R
\docType{class}
\name{gridifyObject-class}
\alias{gridifyObject-class}
\title{gridifyObject class}
\description{
Class for creating an object in a gridify layout.
}
\section{Slots}{

\describe{
\item{\code{row}}{A numeric value, span or sequence specifying the row position of the object.}

\item{\code{col}}{A numeric value, span or sequence specifying the column position of the object.}

\item{\code{height}}{A numeric value specifying the height of the object.}

\item{\code{width}}{A numeric value specifying the width of the object.}

\item{\code{vjust}}{A numeric value in \verb{[0, 1]} specifying the vertical anchoring of the object
within its cell. \code{0} aligns to the bottom, \code{0.5} (default) centers it, \code{1} aligns to the top.}
}}

