% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-classes.R
\docType{class}
\name{gridifyObject-class}
\alias{gridifyObject-class}
\title{gridifyObject class}
\description{
Class for creating an object in a gridify layout.
}
\section{Slots}{

\describe{
\item{\code{row}}{A numeric value, span or sequence specifying the row position of the object.}

\item{\code{col}}{A numeric value, span or sequence specifying the column position of the object.}

\item{\code{height}}{A numeric value specifying the height of the object.}

\item{\code{width}}{A numeric value specifying the width of the object.}
}}

