% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridify-classes.R
\name{gridifyObject}
\alias{gridifyObject}
\title{Create a gridifyObject}
\usage{
gridifyObject(row, col, height = 1, width = 1)
}
\arguments{
\item{row}{A numeric value, span or sequence specifying the row position of the object.}

\item{col}{A numeric value, span or sequence specifying the row position of the object.}

\item{height}{A numeric value specifying the height of the object. Default is 1.}

\item{width}{A numeric value specifying the width of the object. Default is 1.}
}
\value{
An instance of the gridifyObject class.
}
\description{
Function for creating a new instance of the gridifyObject class.
}
\examples{
object <- gridifyObject(row = 1, col = 1, height = 1, width = 1)
}
\seealso{
\code{\link[=gridifyLayout]{gridifyLayout()}}
}
